// frontend2/src/lib/types.ts
export type Product = {
    id: string
    name: string
    brand: string
    price: number
    image: string
    color: string
    description: string
    createdAt: Date
    productType: string
  }

export type Lens = {
    id: string
    name: string
    brand: string
    price: number
    image: string
    description: string
    createdAt: Date
  }